Don't expose the energy expenditure factor
This was exposed to multiply it always by the last distance travelled to get the energy used.
Instead, make move return that energy expenditure of the last distance travelled

+++ core/src/main/java/com/github/thehilikus/alife/agent/motion/RandomWalk.java
@@ -26,7 +26,7 @@ public class RandomWalk extends Legs {
     }
 
     @Override
-    public int move(double speedFactor, SortedSet<ScanResult> scanResults) {
+    public double move(double speedFactor, SortedSet<ScanResult> scanResults) {
         if (shouldTurn()) {
             int originalOrientation = getOrientation();
             turn(Turn.LEFT);
